+++ src/kernel.hpp
@@ -7,11 +7,11 @@
 
 struct Kernel
 {
-    std::vector<double> peaks;
-    double mean;
-    double deviation;
-    double radius;
-    double growthHeight;
+    std::vector<float> peaks;
+    float mean;
+    float deviation;
+    float radius;
+    float growthHeight;
 
     int source;
     int dest;
@@ -23,8 +23,8 @@ struct Kernel
 
     Kernel(
         unsigned int kernelRadius, unsigned int worldWidth, unsigned int worldHeight, unsigned int fourierWidth, unsigned int fourierHeight,
-        const std::vector<double>& kernelRingPeakValues = {1.0}, double growthMean = 0.135, double growthStandardDeviation = 0.015,
-        double relativeRadius = 1.0, double growthHeightScale = 1.0,
+        const std::vector<float>& kernelRingPeakValues = {1.0}, float growthMean = 0.135, float growthStandardDeviation = 0.015,
+        float relativeRadius = 1.0, float growthHeightScale = 1.0,
         int sourceChannel = 0, int destChannel = 0
     );
